+++ apps/dine-pleiepenger/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @navikt/dine-pleiepenger
 
+## 1.9.0
+
+### Minor Changes
+
+- Bruke amplitude fra dekoratøren
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-amplitude@2.28.0
+
 ## 1.8.2
 
 ### Patch Changes