+++ apps/opplaringspenger-soknad/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @navikt/opplaringspenger-soknad
 
+## 1.18.0
+
+### Minor Changes
+
+- Bruke amplitude fra dekoratøren
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/sif-common-amplitude@2.28.0
+
 ## 1.17.1
 
 ### Patch Changes